Move NeuralNetwork class
